requirements.txt
flask==2.2.3
flask_restx==1.1.0
black==23.1.0
pandas==1.5.3
numpy==1.24.2
sentence-transformers==2.2.2
Pillow==9.4.0
requests==2.28.2
wheel==0.38.4